services/ingestion_service/requirements.txt
fastapi
uvicorn
python-multipart

pymupdf
pdfplumber

sentence-transformers
faiss-cpu

numpy
aiofiles
